Do min_frac comparisons as double rather than float
This suffices to make x87 also produce the expected results for
humanely-readable min_frac values such as 0.6.  Fixes #305.

+++ bam2bcf_indel.c
@@ -197,7 +197,7 @@ int bcf_call_gap_prep(int n, int *n_plp, bam_pileup1_t **plp, int pos, bcf_calla
                 j = bam_cigar2qlen(p->b->core.n_cigar, bam_get_cigar(p->b));
                 if (j > max_rd_len) max_rd_len = j;
             }
-            float frac = (float)na/nt;
+            double frac = (double)na/nt;
             if ( !indel_support_ok && na >= bca->min_support && frac >= bca->min_frac )
                 indel_support_ok = 1;
             if ( na > bca->max_support && frac > 0 ) bca->max_support = na, bca->max_frac = frac;
@@ -215,7 +215,7 @@ int bcf_call_gap_prep(int n, int *n_plp, bam_pileup1_t **plp, int pos, bcf_calla
             if (aux[i] != aux[i-1]) ++n_types;
         // Taking totals makes it hard to call rare indels
         if ( !bca->per_sample_flt )
-            indel_support_ok = ( (float)n_alt / n_tot < bca->min_frac || n_alt < bca->min_support ) ? 0 : 1;
+            indel_support_ok = ( (double)n_alt / n_tot < bca->min_frac || n_alt < bca->min_support ) ? 0 : 1;
         if ( n_types == 1 || !indel_support_ok ) { // then skip
             free(aux); return -1;
         }
